+++ src/features/home/Home.tsx
@@ -2,8 +2,8 @@ import { Link } from "react-router-dom";
 import { useWorkspaceStore } from "../../state/stores/workspace-store";
 import { useWorkspaces } from "./workspace-query";
 import { routePaths } from "../../routes/route-paths";
-import { determineDayTime } from "./utils";
 import { useCurrentUser } from "../../data/user/active-user-query";
+import Welcome from "../../components/Welcome";
 
 function Home() {
 	const { setActiveWorkspace } = useWorkspaceStore();
@@ -13,12 +13,7 @@ function Home() {
 	return (
 		<div className="flex flex-col gap-8">
 			<div>
-				<p className="text-[2rem] text-neutral-400">
-					Good {determineDayTime()},{" "}
-					<span className="bg-gradient-to-br from-primary to-primary-light bg-clip-text text-4xl font-medium text-transparent">
-						{currentUser?.full_name}
-					</span>
-				</p>
+				<Welcome name={currentUser?.full_name || ""} />
 				<p className="text-neutral-400">Please select one of the following workspaces</p>
 			</div>
 
@@ -27,7 +22,7 @@ function Home() {
 					<Link
 						to={routePaths.workspaces.detail(item.name)}
 						onClick={() => setActiveWorkspace(item.name)}
-						className="flex w-full select-text flex-col gap-2 rounded-2xl bg-white p-4"
+						className="flex w-full select-text flex-col gap-2 rounded-2xl bg-white from-primary to-primary-light p-4 hover:bg-primary hover:bg-gradient-to-br hover:text-white"
 					>
 						<h2 className="text-xl">{item.name}</h2>
 						{item.description ? <p>{item.description}</p> : <p>No description provided</p>}
